requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.0.0
colorama==0.4.4
configparser==5.2.0
constantly==15.1.0
coverage==6.3.2
coveralls==3.3.1
cryptography==36.0.1
docopt==0.6.2
fire==0.4.0
future==0.18.2
genson==1.2.2
hyperlink==21.0.0
hypothesis==6.39.1
idna==3.3
incremental==21.3.0
iniconfig==1.1.1
joblib==1.1.0
jsonschema==4.4.0
mock==4.0.3
names==0.3.0
nltk==3.7
numpy==1.22.2
packaging==21.3
pandas==1.4.1
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==7.0.1
pytest-cov==3.0.0
python-dateutil==2.8.2
pytz==2021.3
randomname==0.1.5
regex==2022.3.2
requests==2.27.1
rstr==3.1.0
scikit-learn==1.0.2
scipy==1.8.0
simplejson==3.17.6
six==1.16.0
sklearn==0.0
sortedcontainers==2.4.0
tabpy==2.5.0
termcolor==1.1.0
textblob==0.17.1
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.1
tqdm==4.63.0
Twisted==22.2.0
twisted-iocpsupport==1.0.2
typing_extensions==4.1.1
urllib3==1.26.8
zope.interface==5.4.0
